requirements_flask.txt
Flask==3.0.0
python-dotenv==1.0.0
langchain-openai==0.0.5
langchain-core==0.1.23
httpx==0.25.2
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.6.1
